chat-con-datos/requirements.txt
flask
python-dotenv
requests
streamlit 
pandas